+++ game/server/hl2/weapon_stunstick.cpp
@@ -1,4 +1,4 @@
-//========= Copyright © 1996-2005, Valve Corporation, All rights reserved. ============//
+//========= Copyright Valve Corporation, All rights reserved. ============//
 //
 // Purpose:		Stun Stick- beating stick with a zappy end
 //
@@ -211,7 +211,7 @@ void CWeaponStunStick::Operator_HandleAnimEvent( animevent_t *pEvent, CBaseComba
 			VectorMA( pOperator->Weapon_ShootPosition(), 32, vecDirection, vecEnd );
 			// Stretch the swing box down to catch low level physics objects
 			CBaseEntity *pHurt = pOperator->CheckTraceHullAttack( pOperator->Weapon_ShootPosition(), vecEnd, 
-				Vector(-16,-16,-40), Vector(16,16,16), (int)GetDamageForActivity( GetActivity() ), DMG_CLUB, 0.5f, false );
+				Vector(-16,-16,-40), Vector(16,16,16), GetDamageForActivity( GetActivity() ), DMG_CLUB, 0.5f, false );
 			
 			// did I hit someone?
 			if ( pHurt )
